requirements.txt
pygithub
gitpython